Registration and login working as intended
