+++ app/controllers/LoginController.scala
@@ -25,21 +25,6 @@ import scala.util.{Failure, Success}
 @Singleton
 class LoginController @Inject() extends Controller {
 /*
-
-  def checkUserCredentials(username: String, password: String): Boolean = {
-    val user = Customer.findCustomer(username)
-    var status: Boolean = false
-    if (user.password == password) {
-      //log the user in
-      status = true
-    } else {
-      status = false
-    }
-    status
-  }
-
-
-
   def save = Action {
     implicit request =>
       val newLoginForm = LoginForm.bindFromRequest()
@@ -60,20 +45,6 @@ class LoginController @Inject() extends Controller {
           Redirect(routes.HomeController.home()).withSession(customerSession)
       })
   }
-
-  def newLogin = Action {
-    implicit request =>
-      val form = if (request2flash.get("error").isDefined)
-        LoginForm.bind(request2flash.data)
-      else
-        LoginForm
-      Ok(views.html.loginOurs(form))
-  }
-
-  def logout = Action {
-    implicit request =>
-      Redirect(routes.HomeController.home()).withNewSession
-  }
 */
 
  /* private val LoginForm: Form[CustomerLogin] = Form(mapping(
@@ -85,18 +56,14 @@ class LoginController @Inject() extends Controller {
   /** Creates the form and verifies the data **/
    val LoginForm = Form(tuple(
     "Username" -> nonEmptyText.verifying("Username not found!", CustomerDB.findByUsername(_).nonEmpty),
-    "Password" -> nonEmptyText).verifying("user not registered", fa => CustomerDB.checkUserCredentials(fa._1, fa._2)))
+    "Password" -> nonEmptyText).verifying("user not registered", f => checkUserCredentials(f._1, f._2)))
 
   def save = Action {
     implicit request =>
       val loginFormDB = LoginForm.bindFromRequest()
-      loginFormDB.fold(hasErrors = {
-            println("Save method runs here if errors in form")
-        form =>
-          Redirect(routes.LoginController.newLogin()).flashing(Flash(form.data) +
-            ("error" -> Messages("password.error")))
-      }, success = {
+      loginFormDB.fold(success = {
         newLogin =>
+          println("Successful login!")
           val currentCustomer = CustomerDB.findByUsername(newLogin._1).head
           val customerSession = request.session +
             ("customerID" -> currentCustomer.getAs[Int]("customerID").get.toString) +
@@ -106,9 +73,13 @@ class LoginController @Inject() extends Controller {
             ("phone" -> currentCustomer.getAs[String]("phone").get) +
             ("username" -> currentCustomer.getAs[String]("username").get)
           Redirect(routes.HomeController.home()).withSession(customerSession)
+      }, hasErrors = {
+        form =>
+          println("Save method runs here if errors in form")
+          Redirect(routes.LoginController.newLogin()).flashing(Flash(form.data) +
+            ("error" -> Messages("password.error")))
       })
   }
-
 
   def newLogin = Action {
     implicit request =>
@@ -122,5 +93,29 @@ class LoginController @Inject() extends Controller {
   def logout = Action {
     implicit request =>
       Redirect(routes.HomeController.home()).withNewSession
+  }
+
+  /** Match the username and password for login **/
+  def checkUserCredentials(username: String, password: String): Boolean = {
+    var status: Boolean = false
+    val findQuery = BSONDocument(
+      "username" -> username,
+      "password" -> password
+    )
+    val foundUser = MongoConnector.collectionCustomer.find(findQuery).cursor[BSONDocument]().collect[List]()
+    foundUser onComplete {
+      case Failure(e) => throw e
+      case Success(readResult) =>
+        if (readResult.nonEmpty) {
+          status = true
+          println("Status = " + status)
+          println("Current user: " + readResult.head.getAs[String]("username").get)
+        } else {
+          status = false
+        }
+    }
+    Thread.sleep(500)
+    println("End of the method status: " + status)
+    status
   }
 }
