+++ app/controllers/RegistrationController.scala
@@ -12,6 +12,7 @@ import play.api.i18n.Messages.Implicits._
 import play.api.mvc._
 import play.api.mvc.{Action, Controller, Flash, Request}
 import play.api.Play.current
+import scala.concurrent.ExecutionContext.Implicits.global
 
 /**
   * Created by Administrator on 06/07/2016.
@@ -28,48 +29,24 @@ class RegistrationController @Inject() extends Controller {
     "Username" -> nonEmptyText.verifying("Username already exists!", CustomerDB.findByUsername(_).isEmpty),
     "Password" -> nonEmptyText))
 
-  private val userForm: Form[CustomerDetails] =
-    Form(mapping(
-      "First Name" -> nonEmptyText,
-      "Last Name" -> nonEmptyText,
-      "Email" -> nonEmptyText.verifying("validation.email.duplicate", Customer.findByEmail(_).isEmpty),
-      "Telephone" -> nonEmptyText,
-      "Username" -> nonEmptyText.verifying("validation.username.duplicate", Customer.findByUsername(_).isEmpty),
-      "Password" -> nonEmptyText
-    )(CustomerDetails.apply)(CustomerDetails.unapply))
-
- /* def saveCustomer = Action {
-    implicit request =>
-      val newCustomerForm = userForm.bindFromRequest()
-      newCustomerForm.fold(hasErrors = {
-        form =>
-          Redirect(routes.RegistrationController.newCustomer()).flashing(Flash(form.data) + ("error" -> Messages("register.validation.errors")))
-      }, success = {
-        newCustomer =>
-          Customer.add(newCustomer)
-          Redirect(routes.LoginController.newLogin()).flashing("success" -> Messages("customers.new.success", newCustomer.firstName))
-      }
-      )
-  }*/
-
-  /** Save the new customer into the databse **/
-  //TODO - Forms always run with errrors on first submit
+  /** Save the new customer into the database **/
+  //TODO - Forms always run with errors on first submit
   def saveCustomer = Action {
     implicit request =>
       val newCustomerForm = newUserForm.bindFromRequest()
-      newCustomerForm.fold(hasErrors = {
-        form =>
-          println("runs if errors in registration form")
-          Redirect(routes.RegistrationController.newCustomer()).flashing(Flash(form.data) + ("error" -> Messages("register.validation.errors")))
-      }, success = {
+      newCustomerForm.fold(success = {
         newCustomer =>
           //Finds the next customer ID from the database
           val newID = CustomerDB.findNextID()
           val newerCustomer = new CustomerDB(newID, newCustomer._1, newCustomer._2, newCustomer._3, newCustomer._4, newCustomer._5, newCustomer._6, List[CustomerAddressDB](), List[CustomerCardDB]())
-          CustomerDB.CustomerDBWriter.write(newerCustomer)
+          MongoConnector.collectionCustomer.insert(newerCustomer)
           println(newerCustomer.username)
           println("Does it run")
           Redirect(routes.LoginController.newLogin()).flashing("success" -> Messages("customers.new.success"))
+      }, hasErrors = {
+        form =>
+          println("runs if errors in registration form")
+          Redirect(routes.RegistrationController.newCustomer()).flashing(Flash(form.data) + ("error" -> Messages("register.validation.errors")))
       })
   }
