+++ app/models/CustomerDB.scala
@@ -10,7 +10,7 @@ import scala.util.{Failure, Success}
 /**
   * Created by Administrator on 28/07/2016.
   */
-case class CustomerDB(customerID: Int, fName: String, lName: String, email: String, telephone: String, username: String, password: String, addresses: List[CustomerAddressDB], cardDetails: List[CustomerCardDB])
+case class CustomerDB(customerID: Int, fName: String, lName: String, email: String, phone: String, username: String, password: String, addresses: List[CustomerAddressDB], cardDetails: List[CustomerCardDB])
 
 object CustomerDB {
   var userList: List[BSONDocument] = List(BSONDocument())
@@ -22,7 +22,7 @@ object CustomerDB {
         doc.getAs[String]("fName").get,
         doc.getAs[String]("lName").get,
         doc.getAs[String]("email").get,
-        doc.getAs[String]("telephone").get,
+        doc.getAs[String]("phone").get,
         doc.getAs[String]("username").get,
         doc.getAs[String]("password").get,
         doc.getAs[List[CustomerAddressDB]]("addresses").get,
@@ -37,7 +37,7 @@ object CustomerDB {
         "fName" -> customer.fName,
         "lName" -> customer.lName,
         "email" -> customer.email,
-        "telephone" -> customer.telephone,
+        "phone" -> customer.phone,
         "username" -> customer.username,
         "password" -> customer.password,
         "addresses" -> customer.addresses,
@@ -46,7 +46,7 @@ object CustomerDB {
     }
   }
 
-  /** Find by username **/
+  /** Find customer by username **/
   def findByUsername(username: String): List[BSONDocument] = {
     val findQuery = BSONDocument(
       "username" -> username
@@ -57,34 +57,11 @@ object CustomerDB {
       case Success(readResult) =>
         userList = readResult
     }
+    Thread.sleep(500)
     userList
   }
 
-  /** Match the username and password for login **/
-  def checkUserCredentials(username: String, password: String): Boolean = {
-    var status: Boolean = true
-    val findQuery = BSONDocument(
-      "username" -> username,
-      "password" -> password
-    )
-    val foundUser = MongoConnector.collectionCustomer.find(findQuery).cursor[BSONDocument]().collect[List]()
-    foundUser onComplete {
-      case Failure(e) => throw e
-      case Success(readResult) =>
-        if (readResult.nonEmpty) {
-          status = true
-          println("Status = " + status)
-          println("Current user: " + readResult.head.getAs[String]("username").get)
-        } else {
-          status = false
-        }
-    }
-    Thread.sleep(250)
-    println("End of the method status: " + status)
-    status
-  }
-
-  /** Send recovery email **/
+  /** Find customer by email **/
   def findByEmail(email : String): List[BSONDocument] = {
     val findQuery = BSONDocument(
       "email" -> email
@@ -95,6 +72,7 @@ object CustomerDB {
       case Success(readResult) =>
         userList = readResult
     }
+    Thread.sleep(500)
     userList
   }
 
@@ -112,7 +90,7 @@ object CustomerDB {
       case Success(readResult) =>
         nextID = readResult.size + 1
     }
-    Thread.sleep(250)
+    Thread.sleep(500)
     nextID
   }
 }
